pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mike-ai-assistant"
version = "0.1.0"
description = "Local-first personal AI assistant powered by Ollama"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Boss"}
]
keywords = ["ai", "assistant", "ollama", "llm", "local"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "ollama>=0.6.2",
    "rich>=15.0.0",
    "pyyaml>=6.0.3",
    "tiktoken>=0.12.0",
    "ddgs>=9.14.2",
    "chromadb>=1.5.9",
    "python-dotenv>=1.2.2",
    "click>=8.3.3",
    "requests>=2.33.1",
    "prompt-toolkit>=3.0.52",
    "questionary>=2.1.1",
    "pypdf>=6.10.2",
]

[project.optional-dependencies]
ui = [
    "fastapi>=0.136.1",
    "uvicorn>=0.46.0",
    "websockets>=16.0",
    "jinja2>=3.1.6",
    "httpx>=0.28.1",
    "psutil>=6.1.1",
    "python-multipart>=0.0.27",
]
voice = [
    "openai-whisper>=20231117",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
]
telegram = [
    "python-telegram-bot>=20.0",
]
anthropic = [
    "anthropic>=0.40.0",
]
openai = [
    "openai>=1.50.0",
]
gemini = [
    "google-generativeai>=0.8.0",
    "pillow>=10.0.0",
]
providers = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-generativeai>=0.8.0",
    "pillow>=10.0.0",
]
auth = [
    "authlib>=1.7.1",
    "itsdangerous>=2.2.0",
    "argon2-cffi>=25.1.0",
    "python-multipart>=0.0.27",
    "aiosmtplib>=5.1.0",
    "email-validator>=2.3.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.49",
]
all = [
    "mike-ai-assistant[ui,voice,telegram,providers,auth]",
]

[project.scripts]
mike = "mike.cli:main"

[project.urls]
Homepage = "https://github.com/sai21-learn/mike"
Repository = "https://github.com/sai21-learn/mike"
Issues = "https://github.com/sai21-learn/mike/issues"

[tool.hatch.build.targets.wheel]
packages = ["mike"]
